+++ src/lib/components/forms/ControllerTextarea.tsx
@@ -12,7 +12,7 @@ import type {
   FieldValues,
   UseControllerProps,
 } from "react-hook-form";
-import { useController } from "react-hook-form";
+import { useWatch, useController } from "react-hook-form";
 
 import type { TextAreaProps } from "./TextAreaInput";
 
@@ -34,6 +34,11 @@ export const ControllerTextarea = <T extends FieldValues>({
   rules = {},
   ...componentProps
 }: ControllerTextareaProps<T>) => {
+  const watcher = useWatch({
+    name,
+    control,
+  });
+
   const { field } = useController({ name, control, rules });
 
   const isError = !!error;
@@ -54,7 +59,12 @@ export const ControllerTextarea = <T extends FieldValues>({
           {label}
         </FormLabel>
       )}
-      <Textarea resize="none" placeholder={placeholder} />
+      <Textarea
+        resize="none"
+        placeholder={placeholder}
+        value={watcher}
+        onChange={field.onChange}
+      />
       {isError ? (
         <FormErrorMessage className="error-text">{error}</FormErrorMessage>
       ) : (