+++ src/lib/components/modal/contract/index.ts
@@ -1,4 +1,5 @@
 export * from "./AddToOtherList";
-export * from "./EditContract";
+export * from "./EditContractDetails";
 export * from "./RemoveContract";
+export * from "./SaveContractDetails";
 export * from "./SaveNewContract";